Move arbitrary generators to `Test.Ouroboros.Network.Testing.Arbitrary`
This way they are all in one central place.  Now, they are not only used
in `Test.Chain` so it makes sense to have them elsewhere.
